Alternative example
Great really thorough step-by step usage, though may be a little too broken up, making it harder to scan.

My PR might make it a little clearer and easier to get to the meat of the usage when first approaching the package. Also note examples with variables like "a" "b" don't make it *immediately* clear what the structure is or how it should be used. May be obvious as the author, but took me a moment to figure it out.